finished runner and tests
