+++ lib/oxford/network.rb
@@ -1,0 +1,6 @@
+module Oxford
+  class Network < ActiveLdap::Base
+    ldap_mapping :dn_attribute => 'cn', :prefix => "",
+                 :classes => ['top', 'device', 'facterNetwork']
+  end
+end
