+++ lib/oxford/processor.rb
@@ -1,0 +1,6 @@
+module Oxford
+  class Processor < ActiveLdap::Base
+    ldap_mapping :dn_attribute => 'cn', :prefix => "",
+                 :classes => ['top', 'device', 'facterProcessor']
+  end
+end
