+++ lib/oxford/runner.rb
@@ -1,17 +1,16 @@
 module Oxford
   class Runner
+    class << self; attr_accessor :config,:debug end
+    @config = File.join(File.dirname(__FILE__), '../../config/database.yaml')
+    @debug = false
+
     def self.run!
-      debug = false
       Oxford::LDAPAdapter.new
 
       # should run the fact retrievers
       @facts = Oxford::Facts.new
-      begin
-        @host = Oxford::Host.find(@facts.hostname)
-      rescue ActiveLdap::EntryNotFound
-        @host = Oxford::Host.new(@facts.hostname)
-        @host.save
-      end
+      @host = Oxford::Host.new(@facts.hostname)
+      @host.save
 
       # Set facts provided by Facter.
       @facts.all.each do |fact, value|
@@ -19,7 +18,7 @@ module Oxford
         if @host.respond_to?("fact#{f}")
           @host.__send__("fact#{f}=", value.to_s)
         else
-          puts "fact #{f} not being recorded as there is no corresponding entry in the data store" if debug
+          puts "fact #{f} not being recorded as there is no corresponding entry in the data store" if @debug
         end
       end
